refactor(web): migrate data-filter-tester to feature folder
