+++ apps/web/src/tools/data-filter-tester/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { DataFilterTester } from "./ui";
+
+export const tool: ToolModule = { id: "data-filter-tester", Component: DataFilterTester };
