+++ apps/web/src/tools/data-filter-tester/messages.ts
@@ -1,0 +1,20 @@
+import type { LocaleMessages } from "@/tools/types";
+
+export const messages: LocaleMessages = {
+  en: {
+    Tools: {
+      "data-filter-tester": {
+        title: "JSONPath Filter Tester",
+        description: "Run @rfjs/data-filter conditions against sample data live.",
+      },
+    },
+  },
+  "zh-TW": {
+    Tools: {
+      "data-filter-tester": {
+        title: "JSONPath 篩選測試器",
+        description: "對範例資料即時執行 @rfjs/data-filter 條件。",
+      },
+    },
+  },
+};
